requirements.txt
azure-ai-textanalytics==1.0.0
azure-common==1.1.25
azure-core==1.6.0
certifi==2020.4.5.2
chardet==3.0.4
idna==2.9
isodate==0.6.0
msrest==0.6.16
numpy==1.18.5
oauthlib==3.1.0
pandas==1.0.4
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.1.2
requests==2.23.0
requests-oauthlib==1.3.0
six==1.15.0
urllib3==1.25.9
xlrd==1.2.0
